separate each lass in different files
